requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.104.1
h11==0.14.0
idna==3.6
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
motor~=3.3.2
aiohttp~=3.9.1
beautifulsoup4~=4.12.2
validators~=0.22.0
torch~=2.1.2
torchvision~=0.16.2
transformers~=4.36.2
Pillow~=10.1.0